An OnItemClickListener is set in every Activity. Tapping on each word, the pronunciation of the word is played.
A releaseMediaPlayer method is added to every Activity.

An OnCompletionListener is set in every Activity (executing the releaseMediaPlayer method).
After the audio file of the pronunciation of a word is completed, it releases the MediaPlayer in order to free up memory.

The releaseMediaPlayer method is also executed before a new MediaPlayer is created in order to ensure, that the previous one is released.